Moved some code from MelangeHelper to K3DslHelper
