+++ java_execution/java_engine/plugins/org.eclipse.gemoc.execution.sequential.javaengine.ui/src/org/eclipse/gemoc/execution/sequential/javaengine/ui/launcher/tabs/LaunchConfigurationMainTab.java
@@ -32,6 +32,8 @@ import org.eclipse.gemoc.dsl.debug.ide.sirius.ui.launch.AbstractDSLLaunchConfigu
 import org.eclipse.gemoc.execution.sequential.javaengine.PlainK3ExecutionEngine;
 import org.eclipse.gemoc.execution.sequential.javaengine.ui.Activator;
 import org.eclipse.gemoc.execution.sequential.javaengine.ui.launcher.LauncherMessages;
+import org.eclipse.gemoc.executionframework.engine.commons.DslHelper;
+import org.eclipse.gemoc.executionframework.engine.commons.K3DslHelper;
 import org.eclipse.gemoc.executionframework.engine.commons.MelangeHelper;
 import org.eclipse.gemoc.executionframework.engine.ui.commons.RunConfiguration;
 import org.eclipse.gemoc.executionframework.ui.utils.ENamedElementQualifiedNameLabelProvider;
@@ -58,18 +60,6 @@ import org.eclipse.swt.widgets.Text;
 import org.eclipse.ui.PlatformUI;
 import org.eclipse.xtext.naming.DefaultDeclarativeQualifiedNameProvider;
 import org.eclipse.xtext.naming.QualifiedName;
-import org.eclipse.gemoc.commons.eclipse.emf.URIHelper;
-import org.eclipse.gemoc.commons.eclipse.ui.dialogs.SelectAnyIFileDialog;
-import org.eclipse.gemoc.execution.sequential.javaengine.PlainK3ExecutionEngine;
-import org.eclipse.gemoc.execution.sequential.javaengine.ui.Activator;
-import org.eclipse.gemoc.execution.sequential.javaengine.ui.launcher.LauncherMessages;
-import org.eclipse.gemoc.executionframework.engine.commons.DslHelper;
-import org.eclipse.gemoc.executionframework.engine.commons.MelangeHelper;
-import org.eclipse.gemoc.executionframework.engine.ui.commons.RunConfiguration;
-import org.eclipse.gemoc.executionframework.ui.utils.ENamedElementQualifiedNameLabelProvider;
-import org.eclipse.gemoc.xdsmlframework.ui.utils.dialogs.SelectAIRDIFileDialog;
-import org.eclipse.gemoc.xdsmlframework.ui.utils.dialogs.SelectAnyEObjectDialog;
-import org.eclipse.gemoc.xdsmlframework.ui.utils.dialogs.SelectMainMethodDialog;
 import org.osgi.framework.Bundle;
 
 /**
@@ -388,7 +378,7 @@ public class LaunchConfigurationMainTab extends LaunchConfigurationTab {
 					setErrorMessage("Please select a language.");
 				}
 				else{
-					Set<Class<?>> candidateAspects = MelangeHelper.getAspects(_languageCombo.getText());
+					Set<Class<?>> candidateAspects = K3DslHelper.getAspects(_languageCombo.getText());
 					SelectMainMethodDialog dialog = new SelectMainMethodDialog(
 							candidateAspects, new ENamedElementQualifiedNameLabelProvider());
 					int res = dialog.open();
@@ -426,8 +416,8 @@ public class LaunchConfigurationMainTab extends LaunchConfigurationTab {
 							new ENamedElementQualifiedNameLabelProvider()){
 						protected boolean select(EObject obj) {
 							String methodSignature = _entryPointMethodText.getText();
-							String firstParamType = MelangeHelper.getParametersType(methodSignature)[0];
-							String simpleParamType =  MelangeHelper.lastSegment(firstParamType);
+							String firstParamType = K3DslHelper.getParametersTypeFromMethodSignature(methodSignature)[0];
+							String simpleParamType =  K3DslHelper.lastSegment(firstParamType);
 							return obj.eClass().getName().equals(simpleParamType);
 						}
 					};
@@ -560,7 +550,7 @@ public class LaunchConfigurationMainTab extends LaunchConfigurationTab {
 			setErrorMessage(LauncherMessages.SequentialMainTab_Language_not_specified); 
 			return false;
 		}
-		else if(MelangeHelper.getEntryPoints(languageName).isEmpty()){
+		else if(K3DslHelper.getEntryPoints(languageName).isEmpty()){
 			setErrorMessage(LauncherMessages.SequentialMainTab_Language_main_methods_dont_exist); 
 			return false;
 		}
@@ -577,8 +567,8 @@ public class LaunchConfigurationMainTab extends LaunchConfigurationTab {
 			return false;
 		}
 		
-		String[] params =MelangeHelper.getParametersType(mainMethod);
-		String firstParam = MelangeHelper.lastSegment(params[0]);
+		String[] params = K3DslHelper.getParametersTypeFromMethodSignature(mainMethod);
+		String firstParam = DslHelper.lastSegment(params[0]);
 		String rootEClass = getModel().getEObject(rootElement).eClass().getName();
 		if( !(params.length == 1 && firstParam.equals(rootEClass)) ){
 			setErrorMessage(LauncherMessages.SequentialMainTab_Language_incompatible_root_and_main);
